requirements.txt
# Core Framework
streamlit==1.28.1
pandas==2.1.4
numpy==1.24.3

# AI/ML
openai==1.3.0
scikit-learn==1.3.0
langchain==0.1.0
langchain-experimental==0.0.47
langchain-openai==0.0.2

# Data Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Data Processing
python-dotenv==1.0.0
faker==20.1.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
black==23.9.1
flake8==6.1.0
jupyter==1.0.0

# Optional: Advanced features
# xgboost==1.7.6
# lightgbm==4.1.0
